♻️ refactor(api): change layer of oauth-provider service
