+++ apps/api/src/app/auth/mobile/login/[provider]/index.ts
@@ -5,7 +5,7 @@ import {
 	OAUTH_STATE_COOKIE_NAME,
 } from "@/common/constants";
 import { oAuthProviderSchema } from "@/domain/oauth-account/provider";
-import { selectOAuthProviderService } from "@/infrastructure/oauth-provider";
+import { selectOAuthProviderGateway } from "@/interface-adapter/gateway/oauth-provider";
 import { ElysiaWithEnv } from "@/modules/elysia-with-env";
 import { BadRequestException } from "@/modules/error/exceptions";
 import { getAPIBaseUrl, getMobileScheme, validateRedirectUrl } from "@mona-ca/core/utils";
@@ -29,7 +29,7 @@ const Provider = new ElysiaWithEnv({ prefix: "/:provider" })
 			const providerRedirectUrl = new URL(`auth/mobile/login/${provider}/callback`, apiBaseUrl);
 
 			const oAuthUseCase = new OAuthUseCase(
-				selectOAuthProviderService({
+				selectOAuthProviderGateway({
 					provider,
 					env,
 					redirectUrl: providerRedirectUrl.toString(),
