--- apps/api/src/infrastructure/oauth-provider/index.ts
@@ -1,6 +1,0 @@
-export type {
-	IOAuthProviderService as IOAuthProviderGateway,
-	AccountInfo,
-} from "./interface/oauth-provider.service.interface";
-export { DiscordOAuthService } from "./providers/discord.service";
-export { selectOAuthProviderService } from "./utils/select-oauth-provider-service";
