+++ apps/api/src/interface-adapter/gateway/oauth-provider/index.ts
@@ -1,0 +1,6 @@
+export type {
+	IOAuthProviderGateway,
+	AccountInfo,
+} from "./interface/oauth-provider.gateway.interface";
+export { DiscordOAuthGateway } from "./providers/discord.gateway";
+export { selectOAuthProviderGateway } from "./utils/select-oauth-provider-gateway";
