+++ apps/api/src/interface-adapter/gateway/oauth-provider/interface/oauth-provider.gateway.interface.ts
@@ -8,7 +8,7 @@ export type AccountInfo = {
 	emailVerified: boolean;
 };
 
-export interface IOAuthProviderService {
+export interface IOAuthProviderGateway {
 	genAuthUrl(state: string, codeVerifier: string): URL;
 	getTokens(code: string, codeVerifier: string): Promise<OAuth2Tokens>;
 	getAccountInfo(accessToken: string): Promise<AccountInfo | null>;
